Move tlm_key, tlm_keys, TangentLinear, TangentLinearMap, and J_tangent_linears. Add documentation.
